OPE-47 remove testing package from app runtime deps
